Audio:mixin_mixout: move mixin_mixouot header file to module folder
This is a clean up, purpose is de-cluster headers, toml files,
Readme.md etc per module basis, since today everything is scattered
in current code base.
