+++ src/audio/mixin_mixout/mixin_mixout_generic.c
@@ -4,9 +4,10 @@
 //
 // Author: Andrula Song <xiaoyuan.song@intel.com>
 
-#include <ipc4/mixin_mixout.h>
 #include <sof/common.h>
 #include <rtos/string.h>
+
+#include "mixin_mixout.h"
 
 #ifdef MIXIN_MIXOUT_GENERIC
